Implement counts based on obtained statements
