+++ indra/assemblers/html/assembler.py
@@ -16,7 +16,9 @@ from indra.statements import *
 from indra.assemblers.english import EnglishAssembler, AgentWithCoordinates
 from indra.databases import get_identifiers_url
 from indra.util.statement_presentation import group_and_sort_statements, \
-    make_top_level_label_from_names_key, make_stmt_from_sort_key
+    make_top_level_label_from_names_key, make_stmt_from_sort_key, \
+    reader_sources, db_sources, all_sources, get_available_source_counts, \
+    get_available_ev_counts, standardize_counts
 from indra.literature import id_lookup
 
 logger = logging.getLogger(__name__)
@@ -41,23 +43,6 @@ def color_gen(scheme):
     while True:
         for color in color_schemes[scheme]:
             yield color
-
-
-db_sources = ['phosphosite', 'cbn', 'pc11', 'biopax', 'bel_lc',
-              'signor', 'biogrid', 'lincs_drug', 'tas', 'hprd', 'trrust',
-              'ctd', 'virhostnet', 'phosphoelm', 'drugbank', 'omnipath']
-
-reader_sources = ['geneways', 'tees', 'isi', 'trips', 'rlimsp', 'medscan',
-                  'sparser', 'eidos', 'reach']
-
-all_sources = db_sources + reader_sources
-
-# These are mappings where the actual INDRA source, as it appears
-# in the evidence source_api is inconsistent with the colors here and
-# with what comes out of the INDRA DB
-internal_source_mappings = {
-    'bel': 'bel_lc'
-}
 
 
 SOURCE_COLORS = [
@@ -678,40 +663,3 @@ def tag_text(text, tag_info_list):
     # Add the last section of text
     format_text += text[start_pos:]
     return format_text
-
-
-def standardize_counts(counts):
-    """Standardize hash-based counts dicts to be int-keyed."""
-    standardized_counts = {}
-    for k, v in counts.items():
-        try:
-            int_k = int(k)
-            standardized_counts[int_k] = v
-        except ValueError:
-            logger.warning('Could not convert statement hash %s to int' % k)
-    return standardized_counts
-
-
-def get_available_ev_counts(stmts):
-    return {stmt.get_hash(): len(stmt.evidence) for stmt in stmts}
-
-
-def get_available_source_counts(stmts):
-    return {stmt.get_hash(): _get_available_ev_source_counts(stmt.evidence)
-            for stmt in stmts}
-
-
-def _get_available_ev_source_counts(evidences):
-    counts = _get_initial_source_counts()
-    for ev in evidences:
-        sa = internal_source_mappings.get(ev.source_api, ev.source_api)
-        try:
-            counts[sa] += 1
-        except KeyError:
-            continue
-    return counts
-
-
-def _get_initial_source_counts():
-    return {s: 0 for s in all_sources}
-
